+++ Changelog.md
@@ -22,6 +22,7 @@ every new version is a new major version.
 ### Fixed
 
 -   Error loading SVGs introduced in v31.
+-   InlineInput`only emits`onChange`and`onChangeComplete` if value has changed
 
 ## 31 - 2023-04-03